Use local DB to render history page
This commit changes the history page to get data from the local database instance, rather than via the client library. It does this by adding `entry_number` and `previous_entry_number` to the `Entry` model. Three queries are made to the database: one query to get a set of 100 entries ordered descending on key, another query to get the total number of entries and the third to get the required previous entries.