Allow stored register data to be updated
This commit makes use of the new optional `since_entry_number` parameter in the client library when calling `get_records_with_history` and `get_metadata_records_with_history` to specify the point at which to get any new register data from. This allows the frontend to check for new records / entries on a regular basis without needing to recompute all of the stored local data about a particular register.