Add fully dockerised development environment
Defines a non-optimised Docker image with Ruby, node.js, Rails and the
frontend application; and a Docker Compose file with Postgres so the
only dependencies required are Docker and Docker Compose.

The Dockerfile has a hack to overwrite the database configuration so
Rails is able to connect to the dockerised Postgres.

Signed-off-by: Arnau Siches <asiches@gmail.com>