Populate register data via new PostgresDataStore
The updated version of the client library (as yet not merged) allows a data store to be passed into the `get_register` method, giving the caller the option of implementing their own data store which `includes DataStore` module. This commit implements a new `PostgresDataStore`, meaning that no data is stored within the client library but instead directly to the Postgres DB the frontend uses (via temporary, in-memory data structures necessary for calculating correct entries and records).