Change logic to use is_empty_register
Previously the views depended on user_records and system_records scopes.
The view does not need access to system_records and these were also
dependent on a query parameter inserted by the user.

This has been made simpler by creating an explicit variable called
`@is_empty_register`, which is independent of `@records` and is not
affected by user query parameters.