Factor out the link resolver
Attempt to factor out a class to linkify field values.

The original method in the controller included a lot of business
logic about how registers link to each other. I want to extract
this out so the controller doesn't have to know about this.

I've kept in the bit that deals with carinality N fields as this
deals a bit more with presentation.