Add end_date field to Record
Currently, the "end-date" of a record is accessed directly from the
"data" json field, and it's type is an ISO 8601 string.

The "end-date" field name is conventionally used to indicate when a
record stops being accurate. On the registers frontend we use that
to implement a status filter.

This filter should take into account the actual value of the end-date
field, and not just its presence. It needs to be able to handle
partial datetimes, for example "2015" - which we would interpret as
"2015-01-01T00:00:00Z".

In preparation for this, add a field to all records with the expanded
value of the end-date. A record is current if end_date is nil or a
datetime that's in the past.