The archived scope should exclude records ending in the future
Currently we use the presence of an "end-date" value in the data
to indicate that a record is archived.

This is not right, because that value could be a date in the future.

So, change the logic to only make records archived if the end date
is set and in the past.