Only run database migrations on one instance
Today we tried to deploy some changes that included a long
running database migration.

Although the site didn't go down, cloudfoundry said it couldn't
start the app. In the logs we noticed it was throwing
ActiveRecord::ConcurrentMigrationError.

This implements the approach from
https://docs.cloudfoundry.org/buildpacks/ruby/rake-config.html
to ensure migrations only get run on one instance. With this
change the migrations run OK on registers-frontend-research.