Add a new job to redownload a register
Create a job which deletes a registers existing records, nullifies its
root hash, and redownloads the register, inside a transaction.

This allows us to redownload a register from scratch with zero downtime.

We intend to do this in two situations:

- For discovery/alpha registers, we don't enforce the append-only nature
  of the register. So when we go back and change history, we need to
  fetch everything again.

- The java implementation stores system entries seperately, and these
  are not currently retrievable as part of a partial update. So
  as a workaround we need to rebuild the register in registers-frontend.

I've modified the existing PopulateRegisterDataInDbJob job to call
the new job as a fallback if the register is invalid. This is
equivalent to the existing behaviour - which would delete the data
and then retry the job - but it does it inside a transaction.

I've factored out the part that calls the client library into a
new class, and I've removed the `refresh_data` call, because I don't
think we need it. The client library already calls this when you
initialise a register client, so it was attempting to fetch data twice,
which complicated the tests.