Add rake task to force a full redownload of a register
This can be used as a workaround when we break the integrity
of discovery registers, or when we update metadata about a register
(this is not currently retrievable as part of an incremental update)

I've removed the retry mechanism for the local version of
the existing job, because this was to handle the case where
the root hash is invalid. It used to delete the data and raise
and exception, so you had to run the job again to get back to
a working state. In this scenario we now fall back to the
full redownload job and don't raise an exception.