Protect against SQL injection of sort parameter
The sort parameter on the registers page is vulnerable to SQL injection
because we don't validate it before generating the SQL that looks it up
in the JSON object.

I've added a regex to check this at the scope level.  The actual
definition of a field is a bit stricter than this, but the broader regex
is easier to understand, while not allowing arbitrary SQL.