fix: properly ignore the bundle folder created by CI
We're suffering Bundler loosing path to gems on CI which would
indicate that the deploy script (powered by CF) is getting confused by
the artifacts of the CI build (i.e a `.bundle` folder).

This directory should be ignored by CF, however a leading slash is now
guaranteed to work in all configurations so favour `.bundle` instead.