docker: don't run sed to engrain the containerised DB configure
Instead use ENV variables to allow a more generic configuration that
runs both locally and in containers.

This is important when mounting the current directory in the
container (to allow development changes to immediately reflect), since
the database file is then re-read (from the host FS) and needs to
execute well in the container.

Moving the values into ENV means we always get the correct build-time
value of our Docker image.