Revert "BAU: Hijack application rendering to remove cookie banner"
This reverts commit e4817a2a8139f96ec6a8603da0754c77c23d591f.

There is no need for such a hack as `govuk_template` already checks
for `window.GOVUK.addCookieMessage` before rendering the cookie
banner.

The hack led to a rather severe issue where any error (internal 5xx or
not found 404) would fail because:

1. our custom `ErrorsController` tries to render the `:application` layout;
2. which invokes `:remove_cookie_banner`;
3. which is a helper defined on `ApplicationController`;
4. which is *not* inherited by `ErrorsController`.

Which causes the page to crash and our production application to
quickly go extinct if some misguided requests hit it.

[1]: https://github.com/alphagov/govuk_template/blob/f164b2825b1479d7c7d4bdbec4ba9c8e171f504c/source/assets/javascripts/core.js#L5