fix: explicitly set the `seen_cookie_banner` cookie
This inhibits the cookie banner on all pages _but_ the custom error
pages because our `ErrorsController` doesn't inherit
`ApplicationController`. Even if it did, Rails does not support
setting cookies on 404 responses[1] so just let it be.

I think it's fair to deem a tiny cookie banner acceptable on the
occasional error page given that Registers uses the long-deprecated[2]
`govuk-template` to do its rendering and we currently don't have the
resources to consider the upgrade.

[1]: https://github.com/rails/rails/issues/11679
[2]: https://github.com/alphagov/govuk_template/commit/9687d8b53deb72da3e501fe30ed6cc9fcd04dbf7